Tile / Player / There's Not Really a Difference
Tiles now exist for the player to gallavant along. These use unity's tilemap system.

In addition, the Player is made up of tilemaps as well, a functional one and an additional visual one. The player can have fully arbitrary shapes that collide with others in a logical manner.

Transformation is also semi functional. Pressing "Z" transforms one of the players into a block.  This block collides as expected with the player, and even when both active collisions are still sane.

Do keep in mind walls do not work, however.